[LITTLE_FIX]  : main push용 test 파일 추가
