+++ app/app/src/main/res/layout/dialog.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
@@ -7,8 +8,8 @@
 
     <LinearLayout
         android:id="@+id/dialog"
-        android:layout_width="362dp"
-        android:layout_height="533dp"
+        android:layout_width="356dp"
+        android:layout_height="586dp"
         android:background="#000000"
         android:orientation="vertical"
         android:padding="16dp">
@@ -52,38 +53,88 @@
             android:hint="mail"
             android:textColorHint="#979696" />
 
-        <EditText
-            android:id="@+id/event"
+        <androidx.constraintlayout.widget.ConstraintLayout
             android:layout_width="330dp"
-            android:layout_height="50dp"
-            android:layout_gravity="center"
-            android:layout_margin="10dp"
-            android:background="@drawable/border"
-            android:hint="event"
-            android:textColorHint="#979696" />
+            android:layout_height="130dp"
+            android:orientation="horizontal">
+
+            <TextView
+                android:id="@+id/textView"
+                android:layout_width="59dp"
+                android:layout_height="18dp"
+                android:text="Event"
+                android:textColor="#FDFDFD"
+                android:textSize="16sp"
+                android:textStyle="bold"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintTop_toTopOf="parent" />
+
+            <com.google.android.material.chip.Chip
+                android:id="@+id/off"
+                android:layout_width="49dp"
+                android:layout_height="46dp"
+                android:layout_marginStart="40dp"
+                android:layout_marginTop="15dp"
+                android:foregroundGravity="center"
+                android:gravity="center"
+                android:text="Off"
+                android:textColor="@color/black"
+                android:textSize="16sp"
+                android:textStyle="bold"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintTop_toBottomOf="@+id/textView" />
+
+            <com.google.android.material.chip.Chip
+                android:id="@+id/fivemin"
+                android:layout_width="93dp"
+                android:layout_height="46dp"
+                android:layout_marginStart="5dp"
+                android:layout_marginTop="15dp"
+                android:text="5분 후 알림"
+                app:layout_constraintStart_toEndOf="@+id/off"
+                app:layout_constraintTop_toBottomOf="@+id/textView" />
+
+            <com.google.android.material.chip.Chip
+                android:id="@+id/tenmin"
+                android:layout_width="98dp"
+                android:layout_height="46dp"
+                android:layout_marginStart="5dp"
+                android:layout_marginTop="15dp"
+                android:text="10분 후 알림"
+                app:layout_constraintStart_toEndOf="@+id/fivemin"
+                app:layout_constraintTop_toBottomOf="@+id/textView" />
+
+            <com.google.android.material.chip.Chip
+                android:id="@+id/thirtymin"
+                android:layout_width="98dp"
+                android:layout_height="46dp"
+                android:layout_marginStart="40dp"
+                android:text="30분 후 알림"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintTop_toBottomOf="@+id/off" />
+
+        </androidx.constraintlayout.widget.ConstraintLayout>
 
         <LinearLayout
-            android:layout_width="330dp"
-            android:layout_height="61dp"
+            android:layout_width="327dp"
+            android:layout_height="wrap_content"
             android:layout_gravity="center"
             android:layout_margin="15dp"
             android:orientation="horizontal">
 
             <Button
                 android:id="@+id/cancelbtn"
-                android:layout_width="161dp"
-                android:layout_height="50dp"
-                android:layout_weight="1"
+                android:layout_width="160dp"
+                android:layout_height="match_parent"
                 android:backgroundTint="@color/white"
                 android:text="cancel"
                 android:textColor="#000000" />
 
             <Button
                 android:id="@+id/save"
-                android:layout_width="164dp"
-                android:layout_height="50dp"
-                android:layout_marginLeft="10dp"
-                android:layout_weight="1"
+                android:layout_width="160dp"
+                android:layout_height="match_parent"
+                android:layout_marginLeft="5dp"
                 android:backgroundTint="#FFFFFF"
                 android:text="save"
                 android:textColor="#000000" />
